+++ crates/sui-bridge/src/e2e_tests/test_utils.rs
@@ -6,13 +6,23 @@ use crate::abi::EthBridgeConfig;
 use crate::config::default_ed25519_key_pair;
 use crate::crypto::BridgeAuthorityKeyPair;
 use crate::crypto::BridgeAuthorityPublicKeyBytes;
+use crate::crypto::BridgeAuthoritySignInfo;
 use crate::events::*;
 use crate::metrics::BridgeMetrics;
 use crate::server::BridgeNodePublicMetadata;
+use crate::sui_transaction_builder::build_add_tokens_on_sui_transaction;
+use crate::sui_transaction_builder::build_committee_register_transaction;
 use crate::types::BridgeAction;
+use crate::types::BridgeCommitteeValiditySignInfo;
+use crate::types::CertifiedBridgeAction;
+use crate::types::VerifiedCertifiedBridgeAction;
 use crate::utils::get_eth_signer_client;
+use crate::utils::publish_and_register_coins_return_add_coins_on_sui_action;
+use crate::utils::wait_for_server_to_be_up;
 use crate::utils::EthSigner;
 use ethers::types::Address as EthAddress;
+use futures::future::join_all;
+use futures::Future;
 use move_core_types::language_storage::StructTag;
 use prometheus::Registry;
 use rand::rngs::SmallRng;
@@ -28,21 +38,33 @@ use std::path::PathBuf;
 use std::process::Command;
 use std::str::FromStr;
 use std::sync::Arc;
+use sui_json_rpc_api::BridgeReadApiClient;
 use sui_json_rpc_types::SuiEvent;
+use sui_json_rpc_types::SuiExecutionStatus;
+use sui_json_rpc_types::SuiTransactionBlockEffectsAPI;
 use sui_json_rpc_types::SuiTransactionBlockResponse;
 use sui_json_rpc_types::SuiTransactionBlockResponseOptions;
 use sui_json_rpc_types::SuiTransactionBlockResponseQuery;
 use sui_json_rpc_types::TransactionFilter;
 use sui_sdk::wallet_context::WalletContext;
 use sui_test_transaction_builder::TestTransactionBuilder;
+use sui_types::base_types::ObjectID;
+use sui_types::bridge::get_bridge;
+use sui_types::bridge::get_bridge_obj_initial_shared_version;
 use sui_types::bridge::BridgeChainId;
+use sui_types::bridge::BridgeSummary;
+use sui_types::bridge::BridgeTrait;
+use sui_types::bridge::{TOKEN_ID_BTC, TOKEN_ID_ETH, TOKEN_ID_USDC, TOKEN_ID_USDT};
 use sui_types::committee::TOTAL_VOTING_POWER;
 use sui_types::crypto::get_key_pair;
+use sui_types::crypto::ToFromBytes;
 use sui_types::digests::TransactionDigest;
-use sui_types::transaction::{ObjectArg, TransactionData};
+use sui_types::object::Object;
+use sui_types::transaction::{ObjectArg, Transaction, TransactionData};
 use sui_types::SUI_BRIDGE_OBJECT_ID;
 use tokio::join;
 use tokio::task::JoinHandle;
+use tokio::time::Instant;
 
 use tracing::error;
 use tracing::info;
@@ -79,7 +101,7 @@ pub const TEST_PK: &str = "0x4bbbf85ce3377467afe5d46f804f221813b2bb87f24d81f60f1
 /// structs that are needed for testing.
 pub struct BridgeTestCluster {
     pub num_validators: usize,
-    pub test_cluster: TestCluster,
+    pub test_cluster: TestClusterWrapper,
     bridge_client: SuiBridgeClient,
     eth_environment: EthBridgeEnvironment,
     bridge_node_handles: Option<Vec<JoinHandle<()>>>,
@@ -161,8 +183,7 @@ impl BridgeTestClusterBuilder {
             bridge_keys.push(kp.copy());
             bridge_keys_copy.push(kp);
         }
-        let start_cluster_task =
-            tokio::task::spawn(Self::start_test_cluster(bridge_keys, self.num_validators));
+        let start_cluster_task = tokio::task::spawn(Self::start_test_cluster(bridge_keys));
         let start_eth_env_task = tokio::task::spawn(Self::start_eth_env(bridge_keys_copy));
         let (start_cluster_res, start_eth_env_res) = join!(start_cluster_task, start_eth_env_task);
         let test_cluster = start_cluster_res.unwrap();
@@ -179,9 +200,10 @@ impl BridgeTestClusterBuilder {
                     .await,
             );
         }
-        let bridge_client = SuiBridgeClient::new(&test_cluster.fullnode_handle.rpc_url, metrics)
-            .await
-            .unwrap();
+        let bridge_client =
+            SuiBridgeClient::new(&test_cluster.inner.fullnode_handle.rpc_url, metrics)
+                .await
+                .unwrap();
         info!(
             "Bridge committee: {:?}",
             bridge_client
@@ -203,15 +225,11 @@ impl BridgeTestClusterBuilder {
         }
     }
 
-    async fn start_test_cluster(
-        bridge_keys: Vec<BridgeAuthorityKeyPair>,
-        num_validators: usize,
-    ) -> TestCluster {
-        assert_eq!(bridge_keys.len(), num_validators);
-        let test_cluster: test_cluster::TestCluster = TestClusterBuilder::new()
-            .with_num_validators(num_validators)
-            .with_protocol_version(BRIDGE_ENABLE_PROTOCOL_VERSION.into())
-            .build_with_bridge(bridge_keys, true)
+    async fn start_test_cluster(bridge_keys: Vec<BridgeAuthorityKeyPair>) -> TestClusterWrapper {
+        let test_cluster = TestClusterWrapperBuilder::new()
+            .with_bridge_authority_keys(bridge_keys)
+            .with_deploy_tokens(true)
+            .build()
             .await;
         info!("Test cluster built");
         test_cluster
@@ -261,11 +279,11 @@ impl BridgeTestCluster {
     }
 
     pub fn sui_client(&self) -> &SuiClient {
-        &self.test_cluster.fullnode_handle.sui_client
+        &self.test_cluster.inner.fullnode_handle.sui_client
     }
 
     pub fn sui_user_address(&self) -> SuiAddress {
-        self.test_cluster.get_address_0()
+        self.test_cluster.inner.get_address_0()
     }
 
     pub fn sui_chain_id(&self) -> BridgeChainId {
@@ -289,19 +307,19 @@ impl BridgeTestCluster {
     }
 
     pub fn wallet_mut(&mut self) -> &mut WalletContext {
-        self.test_cluster.wallet_mut()
+        self.test_cluster.inner.wallet_mut()
     }
 
     pub fn wallet(&self) -> &WalletContext {
-        &self.test_cluster.wallet
+        &self.test_cluster.inner.wallet
     }
 
     pub fn bridge_authority_key(&self, index: usize) -> BridgeAuthorityKeyPair {
-        self.test_cluster.bridge_authority_keys.as_ref().unwrap()[index].copy()
+        self.test_cluster.bridge_authority_keys[index].copy()
     }
 
     pub fn sui_rpc_url(&self) -> String {
-        self.test_cluster.fullnode_handle.rpc_url.clone()
+        self.test_cluster.inner.fullnode_handle.rpc_url.clone()
     }
 
     pub fn eth_rpc_url(&self) -> String {
@@ -317,6 +335,7 @@ impl BridgeTestCluster {
         sender: SuiAddress,
     ) -> TestTransactionBuilder {
         self.test_cluster
+            .inner
             .test_transaction_builder_with_sender(sender)
             .await
     }
@@ -332,6 +351,7 @@ impl BridgeTestCluster {
         tx_data: &TransactionData,
     ) -> SuiTransactionBlockResponse {
         self.test_cluster
+            .inner
             .sign_and_execute_transaction(tx_data)
             .await
     }
@@ -723,18 +743,16 @@ impl Drop for EthBridgeEnvironment {
 }
 
 pub(crate) async fn start_bridge_cluster(
-    test_cluster: &TestCluster,
+    test_cluster: &TestClusterWrapper,
     eth_environment: &EthBridgeEnvironment,
     approved_governance_actions: Vec<Vec<BridgeAction>>,
 ) -> Vec<JoinHandle<()>> {
     let bridge_authority_keys = test_cluster
         .bridge_authority_keys
-        .as_ref()
-        .unwrap()
         .iter()
         .map(|k| k.copy())
         .collect::<Vec<_>>();
-    let bridge_server_ports = test_cluster.bridge_server_ports.as_ref().unwrap();
+    let bridge_server_ports = test_cluster.bridge_server_ports.clone();
     assert_eq!(bridge_authority_keys.len(), bridge_server_ports.len());
     assert_eq!(
         bridge_authority_keys.len(),
@@ -778,7 +796,7 @@ pub(crate) async fn start_bridge_cluster(
                 eth_contracts_start_block_override: None,
             },
             sui: SuiConfig {
-                sui_rpc_url: test_cluster.fullnode_handle.rpc_url.clone(),
+                sui_rpc_url: test_cluster.inner.fullnode_handle.rpc_url.clone(),
                 sui_bridge_chain_id: BridgeChainId::SuiCustom as u8,
                 bridge_client_key_path: None,
                 bridge_client_gas_object: None,
@@ -849,4 +867,346 @@ impl Drop for TempDir {
             eprintln!("Failed to remove temp dir: {:?}", e);
         }
     }
+}
+
+pub struct TestClusterWrapperBuilder {
+    protocol_version: u64,
+    bridge_authority_keys: Vec<BridgeAuthorityKeyPair>,
+    deploy_tokens: bool,
+}
+
+impl TestClusterWrapperBuilder {
+    pub fn new() -> Self {
+        Self {
+            protocol_version: BRIDGE_ENABLE_PROTOCOL_VERSION,
+            bridge_authority_keys: vec![],
+            deploy_tokens: false,
+        }
+    }
+
+    pub fn with_protocol_version(mut self, version: u64) -> Self {
+        self.protocol_version = version;
+        self
+    }
+
+    pub fn with_bridge_authority_keys(mut self, keys: Vec<BridgeAuthorityKeyPair>) -> Self {
+        self.bridge_authority_keys = keys;
+        self
+    }
+
+    pub fn with_deploy_tokens(mut self, deploy_tokens: bool) -> Self {
+        self.deploy_tokens = deploy_tokens;
+        self
+    }
+
+    pub async fn build(self) -> TestClusterWrapper {
+        assert_ne!(self.bridge_authority_keys.len(), 0);
+        let num_validators = self.bridge_authority_keys.len();
+        let builder = TestClusterBuilder::new().with_protocol_version(self.protocol_version.into());
+
+        let timer = Instant::now();
+        let gas_objects_for_authority_keys = self
+            .bridge_authority_keys
+            .iter()
+            .map(|k| {
+                let address = SuiAddress::from(k.public());
+                Object::with_id_owner_for_testing(ObjectID::random(), address)
+            })
+            .collect::<Vec<_>>();
+        let mut test_cluster = builder
+            .with_num_validators(num_validators)
+            .with_objects(gas_objects_for_authority_keys)
+            .build()
+            .await;
+        info!(
+            "TestCluster build took {:?} secs",
+            timer.elapsed().as_secs()
+        );
+        let ref_gas_price = test_cluster.get_reference_gas_price().await;
+        let bridge_arg = get_mut_bridge_arg(&test_cluster).await.unwrap();
+        assert_eq!(
+            self.bridge_authority_keys.len(),
+            test_cluster.swarm.active_validators().count()
+        );
+
+        // Committee registers themselves
+        let mut server_ports = vec![];
+        let mut tasks = vec![];
+        let quorum_driver_api = test_cluster.quorum_driver_api().clone();
+        // Reorder the nodes so that the last node has the largest stake.
+        let validator_with_max_stake = test_cluster
+            .sui_client()
+            .governance_api()
+            .get_committee_info(None)
+            .await
+            .unwrap()
+            .validators
+            .iter()
+            .max_by(|a, b| a.0.cmp(&b.0))
+            .unwrap()
+            .0;
+        let node_with_max_stake = test_cluster
+            .swarm
+            .active_validators()
+            .find(|v| v.config().protocol_public_key() == validator_with_max_stake)
+            .unwrap();
+        let other_nodes = test_cluster
+            .swarm
+            .active_validators()
+            .filter(|v| v.config().protocol_public_key() != validator_with_max_stake)
+            .collect::<Vec<_>>();
+        let reordered_nodes = other_nodes
+            .iter()
+            .chain(std::iter::once(&node_with_max_stake));
+        for (node, kp) in reordered_nodes.zip(self.bridge_authority_keys.iter()) {
+            let validator_address = node.config().sui_address();
+            // create committee registration tx
+            let gas = test_cluster
+                .wallet
+                .get_one_gas_object_owned_by_address(validator_address)
+                .await
+                .unwrap()
+                .unwrap();
+
+            let server_port = get_available_port("127.0.0.1");
+            let server_url = format!("http://127.0.0.1:{}", server_port);
+            server_ports.push(server_port);
+            let data = build_committee_register_transaction(
+                validator_address,
+                &gas,
+                bridge_arg,
+                kp.public().as_bytes().to_vec(),
+                &server_url,
+                ref_gas_price,
+                1000000000,
+            )
+            .unwrap();
+
+            let tx = Transaction::from_data_and_signer(
+                data,
+                vec![node.config().account_key_pair.keypair()],
+            );
+            let api_clone = quorum_driver_api.clone();
+            tasks.push(async move {
+                api_clone
+                    .execute_transaction_block(
+                        tx,
+                        SuiTransactionBlockResponseOptions::new().with_effects(),
+                        None,
+                    )
+                    .await
+            });
+        }
+
+        if self.deploy_tokens {
+            let timer = Instant::now();
+            let token_ids = vec![TOKEN_ID_BTC, TOKEN_ID_ETH, TOKEN_ID_USDC, TOKEN_ID_USDT];
+            let token_prices = vec![500_000_000u64, 30_000_000u64, 1_000u64, 1_000u64];
+            let action = publish_and_register_coins_return_add_coins_on_sui_action(
+                test_cluster.wallet(),
+                bridge_arg,
+                vec![
+                    Path::new("../../bridge/move/tokens/btc").into(),
+                    Path::new("../../bridge/move/tokens/eth").into(),
+                    Path::new("../../bridge/move/tokens/usdc").into(),
+                    Path::new("../../bridge/move/tokens/usdt").into(),
+                ],
+                token_ids,
+                token_prices,
+                0,
+            );
+            let action = action.await;
+            info!("register tokens took {:?} secs", timer.elapsed().as_secs());
+            let sig_map = self
+                .bridge_authority_keys
+                .iter()
+                .map(|key| {
+                    (
+                        key.public().into(),
+                        BridgeAuthoritySignInfo::new(&action, key).signature,
+                    )
+                })
+                .collect::<BTreeMap<_, _>>();
+            let certified_action = CertifiedBridgeAction::new_from_data_and_sig(
+                action,
+                BridgeCommitteeValiditySignInfo {
+                    signatures: sig_map.clone(),
+                },
+            );
+            let verifired_action_cert =
+                VerifiedCertifiedBridgeAction::new_from_verified(certified_action);
+            let sender_address = test_cluster.get_address_0();
+
+            await_committee_register_tasks(&test_cluster, tasks).await;
+
+            // Wait until committee is set up
+            trigger_reconfiguration_if_not_yet_and_assert_bridge_committee_initialized(
+                &test_cluster,
+            )
+            .await;
+
+            let tx = build_add_tokens_on_sui_transaction(
+                sender_address,
+                &test_cluster
+                    .wallet
+                    .get_one_gas_object_owned_by_address(sender_address)
+                    .await
+                    .unwrap()
+                    .unwrap(),
+                verifired_action_cert,
+                bridge_arg,
+                ref_gas_price,
+            )
+            .unwrap();
+
+            let response = test_cluster.sign_and_execute_transaction(&tx).await;
+            assert_eq!(
+                response.effects.unwrap().status(),
+                &SuiExecutionStatus::Success
+            );
+            info!("Deploy tokens took {:?} secs", timer.elapsed().as_secs());
+        } else {
+            await_committee_register_tasks(&test_cluster, tasks).await;
+        }
+        async fn await_committee_register_tasks(
+            test_cluster: &TestCluster,
+            tasks: Vec<
+                impl Future<Output = Result<SuiTransactionBlockResponse, sui_sdk::error::Error>>,
+            >,
+        ) {
+            // The tx may fail if a member tries to register when the committee is already finalized.
+            // In that case, we just need to check the committee members is not empty since once
+            // the committee is finalized, it should not be empty.
+            let responses = join_all(tasks).await;
+            let mut has_failure = false;
+            for response in responses {
+                if response.unwrap().effects.unwrap().status() != &SuiExecutionStatus::Success {
+                    has_failure = true;
+                }
+            }
+            if has_failure {
+                let bridge_summary = get_bridge_summary(test_cluster).await;
+                assert_ne!(bridge_summary.committee.members.len(), 0);
+            }
+        }
+
+        info!(
+            "TestCluster build_with_bridge took {:?} secs",
+            timer.elapsed().as_secs()
+        );
+        TestClusterWrapper {
+            inner: test_cluster,
+            bridge_authority_keys: self.bridge_authority_keys,
+            bridge_server_ports: server_ports,
+        }
+    }
+}
+
+impl Default for TestClusterWrapperBuilder {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+pub struct TestClusterWrapper {
+    pub inner: TestCluster,
+    pub bridge_authority_keys: Vec<BridgeAuthorityKeyPair>,
+    pub bridge_server_ports: Vec<u16>,
+}
+
+impl TestClusterWrapper {
+    pub fn authority_keys_clone(&self) -> Vec<BridgeAuthorityKeyPair> {
+        self.bridge_authority_keys
+            .iter()
+            .map(|k| k.copy())
+            .collect()
+    }
+
+    pub async fn trigger_reconfiguration_if_not_yet_and_assert_bridge_committee_initialized(&self) {
+        trigger_reconfiguration_if_not_yet_and_assert_bridge_committee_initialized(&self.inner)
+            .await
+    }
+
+    // Wait for bridge node in the cluster to be up and running.
+    pub async fn wait_for_bridge_cluster_to_be_up(&self, timeout_sec: u64) {
+        let bridge_ports = self.bridge_server_ports.clone();
+        let mut tasks = vec![];
+        for port in bridge_ports.iter() {
+            let server_url = format!("http://127.0.0.1:{}", port);
+            tasks.push(wait_for_server_to_be_up(server_url, timeout_sec));
+        }
+        join_all(tasks)
+            .await
+            .into_iter()
+            .collect::<anyhow::Result<Vec<_>>>()
+            .unwrap();
+    }
+
+    pub async fn get_mut_bridge_arg(&self) -> Option<ObjectArg> {
+        get_mut_bridge_arg(&self.inner).await
+    }
+
+    pub async fn get_bridge_summary(&self) -> BridgeSummary {
+        get_bridge_summary(&self.inner).await
+    }
+}
+
+async fn get_bridge_summary(test_cluster: &TestCluster) -> BridgeSummary {
+    test_cluster
+        .sui_client()
+        .http()
+        .get_latest_bridge()
+        .await
+        .unwrap()
+}
+
+async fn get_mut_bridge_arg(test_cluster: &TestCluster) -> Option<ObjectArg> {
+    get_bridge_obj_initial_shared_version(
+        test_cluster
+            .fullnode_handle
+            .sui_node
+            .state()
+            .get_object_store(),
+    )
+    .unwrap()
+    .map(|seq| ObjectArg::SharedObject {
+        id: SUI_BRIDGE_OBJECT_ID,
+        initial_shared_version: seq,
+        mutable: true,
+    })
+}
+
+async fn trigger_reconfiguration_if_not_yet_and_assert_bridge_committee_initialized(
+    test_cluster: &TestCluster,
+) {
+    let mut bridge = get_bridge(
+        test_cluster
+            .fullnode_handle
+            .sui_node
+            .state()
+            .get_object_store(),
+    )
+    .unwrap();
+    if !bridge.committee().members.contents.is_empty() {
+        assert_eq!(
+            test_cluster.swarm.active_validators().count(),
+            bridge.committee().members.contents.len()
+        );
+        return;
+    }
+    // wait for next epoch
+    test_cluster.trigger_reconfiguration().await;
+    bridge = get_bridge(
+        test_cluster
+            .fullnode_handle
+            .sui_node
+            .state()
+            .get_object_store(),
+    )
+    .unwrap();
+    // Committee should be initiated
+    assert!(bridge.committee().member_registrations.contents.is_empty());
+    assert_eq!(
+        test_cluster.swarm.active_validators().count(),
+        bridge.committee().members.contents.len()
+    );
 }
