--- crates/sui-e2e-tests/tests/bridge_tests.rs
@@ -1,105 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-use sui_bridge::crypto::BridgeAuthorityKeyPair;
-use sui_bridge::BRIDGE_ENABLE_PROTOCOL_VERSION;
-use sui_json_rpc_api::BridgeReadApiClient;
-use sui_macros::sim_test;
-use sui_types::bridge::get_bridge;
-use sui_types::bridge::BridgeTrait;
-use sui_types::crypto::get_key_pair;
-use sui_types::SUI_BRIDGE_OBJECT_ID;
-use test_cluster::TestClusterBuilder;
-
-#[sim_test]
-async fn test_create_bridge_state_object() {
-    let test_cluster = TestClusterBuilder::new()
-        .with_protocol_version((BRIDGE_ENABLE_PROTOCOL_VERSION - 1).into())
-        .with_epoch_duration_ms(20000)
-        .build()
-        .await;
-
-    let handles = test_cluster.all_node_handles();
-
-    // no node has the bridge state object yet
-    for h in &handles {
-        h.with(|node| {
-            assert!(node
-                .state()
-                .get_object_cache_reader()
-                .get_latest_object_ref_or_tombstone(SUI_BRIDGE_OBJECT_ID)
-                .unwrap()
-                .is_none());
-        });
-    }
-
-    // wait until feature is enabled
-    test_cluster
-        .wait_for_protocol_version(BRIDGE_ENABLE_PROTOCOL_VERSION.into())
-        .await;
-    // wait until next epoch - authenticator state object is created at the end of the first epoch
-    // in which it is supported.
-    test_cluster.wait_for_epoch_all_nodes(2).await; // protocol upgrade completes in epoch 1
-
-    for h in &handles {
-        h.with(|node| {
-            node.state()
-                .get_object_cache_reader()
-                .get_latest_object_ref_or_tombstone(SUI_BRIDGE_OBJECT_ID)
-                .unwrap()
-                .expect("auth state object should exist");
-        });
-    }
-}
-
-#[tokio::test]
-async fn test_committee_registration() {
-    telemetry_subscribers::init_for_testing();
-    let mut bridge_keys = vec![];
-    for _ in 0..=3 {
-        let (_, kp): (_, BridgeAuthorityKeyPair) = get_key_pair();
-        bridge_keys.push(kp);
-    }
-    let test_cluster: test_cluster::TestCluster = TestClusterBuilder::new()
-        .with_protocol_version((BRIDGE_ENABLE_PROTOCOL_VERSION).into())
-        .build_with_bridge(bridge_keys, false)
-        .await;
-
-    let bridge = get_bridge(
-        test_cluster
-            .fullnode_handle
-            .sui_node
-            .state()
-            .get_object_store(),
-    )
-    .unwrap();
-
-    // Member should be empty before end of epoch
-    assert!(bridge.committee().members.contents.is_empty());
-    assert_eq!(
-        test_cluster.swarm.active_validators().count(),
-        bridge.committee().member_registrations.contents.len()
-    );
-
-    test_cluster
-        .trigger_reconfiguration_if_not_yet_and_assert_bridge_committee_initialized()
-        .await;
-}
-
-#[tokio::test]
-async fn test_bridge_api_compatibility() {
-    let test_cluster: test_cluster::TestCluster = TestClusterBuilder::new()
-        .with_protocol_version(BRIDGE_ENABLE_PROTOCOL_VERSION.into())
-        .build()
-        .await;
-
-    test_cluster.trigger_reconfiguration().await;
-    let client = test_cluster.rpc_client();
-    client.get_latest_bridge().await.unwrap();
-    // TODO: assert fields in summary
-
-    client
-        .get_bridge_object_initial_shared_version()
-        .await
-        .unwrap();
-}
