+++ crates/test-cluster/Cargo.toml
@@ -31,7 +31,6 @@ sui-swarm.workspace = true
 sui-types = { workspace = true, features = ["test-utils"] }
 prometheus.workspace = true
 sui-keys.workspace = true
-sui-bridge.workspace = true
 sui-sdk.workspace = true
 sui-test-transaction-builder.workspace = true
