+++ crates/test-cluster/src/lib.rs
@@ -1,35 +1,21 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use futures::Future;
 use futures::{future::join_all, StreamExt};
-use jsonrpsee::core::RpcResult;
 use jsonrpsee::http_client::{HttpClient, HttpClientBuilder};
 use rand::{distributions::*, rngs::OsRng, seq::SliceRandom};
-use std::collections::BTreeMap;
 use std::collections::HashMap;
 use std::net::SocketAddr;
 use std::num::NonZeroUsize;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::sync::{Arc, Mutex};
 use std::time::Duration;
-use sui_bridge::crypto::{BridgeAuthorityKeyPair, BridgeAuthoritySignInfo};
-use sui_bridge::sui_transaction_builder::build_add_tokens_on_sui_transaction;
-use sui_bridge::sui_transaction_builder::build_committee_register_transaction;
-use sui_bridge::types::BridgeCommitteeValiditySignInfo;
-use sui_bridge::types::CertifiedBridgeAction;
-use sui_bridge::types::VerifiedCertifiedBridgeAction;
-use sui_bridge::utils::publish_and_register_coins_return_add_coins_on_sui_action;
-use sui_bridge::utils::wait_for_server_to_be_up;
 use sui_config::genesis::Genesis;
-use sui_config::local_ip_utils::get_available_port;
 use sui_config::node::{AuthorityOverloadConfig, DBCheckpointConfig, RunWithRange};
 use sui_config::{Config, SUI_CLIENT_CONFIG, SUI_NETWORK_CONFIG};
 use sui_config::{NodeConfig, PersistedConfig, SUI_KEYSTORE_FILENAME};
 use sui_core::authority_aggregator::AuthorityAggregator;
 use sui_core::authority_client::NetworkAuthorityClient;
-use sui_json_rpc_api::BridgeReadApiClient;
-use sui_json_rpc_types::SuiTransactionBlockResponseOptions;
 use sui_json_rpc_types::{
     SuiExecutionStatus, SuiTransactionBlockEffectsAPI, SuiTransactionBlockResponse,
     TransactionFilter,
@@ -54,12 +40,10 @@ use sui_swarm_config::node_config_builder::{FullnodeConfigBuilder, ValidatorConf
 use sui_test_transaction_builder::TestTransactionBuilder;
 use sui_types::base_types::ConciseableName;
 use sui_types::base_types::{AuthorityName, ObjectID, ObjectRef, SuiAddress};
-use sui_types::bridge::{get_bridge, TOKEN_ID_BTC, TOKEN_ID_ETH, TOKEN_ID_USDC, TOKEN_ID_USDT};
-use sui_types::bridge::{get_bridge_obj_initial_shared_version, BridgeSummary, BridgeTrait};
 use sui_types::committee::CommitteeTrait;
 use sui_types::committee::{Committee, EpochId};
+use sui_types::crypto::KeypairTraits;
 use sui_types::crypto::SuiKeyPair;
-use sui_types::crypto::{KeypairTraits, ToFromBytes};
 use sui_types::effects::{TransactionEffects, TransactionEvents};
 use sui_types::error::SuiResult;
 use sui_types::governance::MIN_VALIDATOR_JOINING_STAKE_MIST;
@@ -71,10 +55,8 @@ use sui_types::sui_system_state::SuiSystemStateTrait;
 use sui_types::supported_protocol_versions::SupportedProtocolVersions;
 use sui_types::traffic_control::{PolicyConfig, RemoteFirewallConfig};
 use sui_types::transaction::{
-    CertifiedTransaction, ObjectArg, Transaction, TransactionData, TransactionDataAPI,
-    TransactionKind,
+    CertifiedTransaction, Transaction, TransactionData, TransactionDataAPI, TransactionKind,
 };
-use sui_types::SUI_BRIDGE_OBJECT_ID;
 use tokio::time::{timeout, Instant};
 use tokio::{task::JoinHandle, time::sleep};
 use tracing::{error, info};
@@ -108,9 +90,6 @@ pub struct TestCluster {
     pub swarm: Swarm,
     pub wallet: WalletContext,
     pub fullnode_handle: FullNodeHandle,
-
-    pub bridge_authority_keys: Option<Vec<BridgeAuthorityKeyPair>>,
-    pub bridge_server_ports: Option<Vec<u16>>,
 }
 
 impl TestCluster {
@@ -266,10 +245,6 @@ impl TestCluster {
             .await
             .unwrap()
             .compute_object_reference()
-    }
-
-    pub async fn get_bridge_summary(&self) -> RpcResult<BridgeSummary> {
-        self.sui_client().http().get_latest_bridge().await
     }
 
     pub async fn get_object_or_tombstone_from_fullnode_store(
@@ -514,54 +489,6 @@ impl TestCluster {
             })
             .await;
         }
-    }
-
-    pub async fn trigger_reconfiguration_if_not_yet_and_assert_bridge_committee_initialized(&self) {
-        let mut bridge =
-            get_bridge(self.fullnode_handle.sui_node.state().get_object_store()).unwrap();
-        if !bridge.committee().members.contents.is_empty() {
-            assert_eq!(
-                self.swarm.active_validators().count(),
-                bridge.committee().members.contents.len()
-            );
-            return;
-        }
-        // wait for next epoch
-        self.trigger_reconfiguration().await;
-        bridge = get_bridge(self.fullnode_handle.sui_node.state().get_object_store()).unwrap();
-        // Committee should be initiated
-        assert!(bridge.committee().member_registrations.contents.is_empty());
-        assert_eq!(
-            self.swarm.active_validators().count(),
-            bridge.committee().members.contents.len()
-        );
-    }
-
-    // Wait for bridge node in the cluster to be up and running.
-    pub async fn wait_for_bridge_cluster_to_be_up(&self, timeout_sec: u64) {
-        let bridge_ports = self.bridge_server_ports.as_ref().unwrap();
-        let mut tasks = vec![];
-        for port in bridge_ports.iter() {
-            let server_url = format!("http://127.0.0.1:{}", port);
-            tasks.push(wait_for_server_to_be_up(server_url, timeout_sec));
-        }
-        join_all(tasks)
-            .await
-            .into_iter()
-            .collect::<anyhow::Result<Vec<_>>>()
-            .unwrap();
-    }
-
-    pub async fn get_mut_bridge_arg(&self) -> Option<ObjectArg> {
-        get_bridge_obj_initial_shared_version(
-            self.fullnode_handle.sui_node.state().get_object_store(),
-        )
-        .unwrap()
-        .map(|seq| ObjectArg::SharedObject {
-            id: SUI_BRIDGE_OBJECT_ID,
-            initial_shared_version: seq,
-            mutable: true,
-        })
     }
 
     pub async fn wait_for_authenticator_state_update(&self) {
@@ -1191,210 +1118,7 @@ impl TestClusterBuilder {
             swarm,
             wallet,
             fullnode_handle,
-            bridge_authority_keys: None,
-            bridge_server_ports: None,
         }
-    }
-
-    pub async fn build_with_bridge(
-        self,
-        // Note: caller should make sure to keep the authority with largest stake at the end of the list.
-        // This is because we try to set up eth env and sui cluster at the same time. Since we use evenly
-        // distributed stake, we want to keep the authority with the remainder consistent in eth contracts
-        // and sui cluster.
-        bridge_authority_keys: Vec<BridgeAuthorityKeyPair>,
-        deploy_tokens: bool,
-    ) -> TestCluster {
-        let timer = Instant::now();
-        let gas_objects_for_authority_keys = bridge_authority_keys
-            .iter()
-            .map(|k| {
-                let address = SuiAddress::from(k.public());
-                Object::with_id_owner_for_testing(ObjectID::random(), address)
-            })
-            .collect::<Vec<_>>();
-        let mut test_cluster = self
-            .with_objects(gas_objects_for_authority_keys)
-            .build()
-            .await;
-        info!(
-            "TestCluster build took {:?} secs",
-            timer.elapsed().as_secs()
-        );
-        let ref_gas_price = test_cluster.get_reference_gas_price().await;
-        let bridge_arg = test_cluster.get_mut_bridge_arg().await.unwrap();
-        assert_eq!(
-            bridge_authority_keys.len(),
-            test_cluster.swarm.active_validators().count()
-        );
-
-        // Committee registers themselves
-        let mut server_ports = vec![];
-        let mut tasks = vec![];
-        let quorum_driver_api = test_cluster.quorum_driver_api().clone();
-        // Reorder the nodes so that the last node has the largest stake.
-        let validator_with_max_stake = test_cluster
-            .sui_client()
-            .governance_api()
-            .get_committee_info(None)
-            .await
-            .unwrap()
-            .validators
-            .iter()
-            .max_by(|a, b| a.0.cmp(&b.0))
-            .unwrap()
-            .0;
-        let node_with_max_stake = test_cluster
-            .swarm
-            .active_validators()
-            .find(|v| v.config().protocol_public_key() == validator_with_max_stake)
-            .unwrap();
-        let other_nodes = test_cluster
-            .swarm
-            .active_validators()
-            .filter(|v| v.config().protocol_public_key() != validator_with_max_stake)
-            .collect::<Vec<_>>();
-        let reordered_nodes = other_nodes
-            .iter()
-            .chain(std::iter::once(&node_with_max_stake));
-        for (node, kp) in reordered_nodes.zip(bridge_authority_keys.iter()) {
-            let validator_address = node.config().sui_address();
-            // create committee registration tx
-            let gas = test_cluster
-                .wallet
-                .get_one_gas_object_owned_by_address(validator_address)
-                .await
-                .unwrap()
-                .unwrap();
-
-            let server_port = get_available_port("127.0.0.1");
-            let server_url = format!("http://127.0.0.1:{}", server_port);
-            server_ports.push(server_port);
-            let data = build_committee_register_transaction(
-                validator_address,
-                &gas,
-                bridge_arg,
-                kp.public().as_bytes().to_vec(),
-                &server_url,
-                ref_gas_price,
-                1000000000,
-            )
-            .unwrap();
-
-            let tx = Transaction::from_data_and_signer(
-                data,
-                vec![node.config().account_key_pair.keypair()],
-            );
-            let api_clone = quorum_driver_api.clone();
-            tasks.push(async move {
-                api_clone
-                    .execute_transaction_block(
-                        tx,
-                        SuiTransactionBlockResponseOptions::new().with_effects(),
-                        None,
-                    )
-                    .await
-            });
-        }
-
-        if deploy_tokens {
-            let timer = Instant::now();
-            let token_ids = vec![TOKEN_ID_BTC, TOKEN_ID_ETH, TOKEN_ID_USDC, TOKEN_ID_USDT];
-            let token_prices = vec![500_000_000u64, 30_000_000u64, 1_000u64, 1_000u64];
-            let action = publish_and_register_coins_return_add_coins_on_sui_action(
-                test_cluster.wallet(),
-                bridge_arg,
-                vec![
-                    Path::new("../../bridge/move/tokens/btc").into(),
-                    Path::new("../../bridge/move/tokens/eth").into(),
-                    Path::new("../../bridge/move/tokens/usdc").into(),
-                    Path::new("../../bridge/move/tokens/usdt").into(),
-                ],
-                token_ids,
-                token_prices,
-                0,
-            );
-            let action = action.await;
-            info!("register tokens took {:?} secs", timer.elapsed().as_secs());
-            let sig_map = bridge_authority_keys
-                .iter()
-                .map(|key| {
-                    (
-                        key.public().into(),
-                        BridgeAuthoritySignInfo::new(&action, key).signature,
-                    )
-                })
-                .collect::<BTreeMap<_, _>>();
-            let certified_action = CertifiedBridgeAction::new_from_data_and_sig(
-                action,
-                BridgeCommitteeValiditySignInfo {
-                    signatures: sig_map.clone(),
-                },
-            );
-            let verifired_action_cert =
-                VerifiedCertifiedBridgeAction::new_from_verified(certified_action);
-            let sender_address = test_cluster.get_address_0();
-
-            await_committee_register_tasks(&test_cluster, tasks).await;
-
-            // Wait until committee is set up
-            test_cluster
-                .trigger_reconfiguration_if_not_yet_and_assert_bridge_committee_initialized()
-                .await;
-
-            let tx = build_add_tokens_on_sui_transaction(
-                sender_address,
-                &test_cluster
-                    .wallet
-                    .get_one_gas_object_owned_by_address(sender_address)
-                    .await
-                    .unwrap()
-                    .unwrap(),
-                verifired_action_cert,
-                bridge_arg,
-                ref_gas_price,
-            )
-            .unwrap();
-
-            let response = test_cluster.sign_and_execute_transaction(&tx).await;
-            assert_eq!(
-                response.effects.unwrap().status(),
-                &SuiExecutionStatus::Success
-            );
-            info!("Deploy tokens took {:?} secs", timer.elapsed().as_secs());
-        } else {
-            await_committee_register_tasks(&test_cluster, tasks).await;
-        }
-
-        async fn await_committee_register_tasks(
-            test_cluster: &TestCluster,
-            tasks: Vec<
-                impl Future<Output = Result<SuiTransactionBlockResponse, sui_sdk::error::Error>>,
-            >,
-        ) {
-            // The tx may fail if a member tries to register when the committee is already finalized.
-            // In that case, we just need to check the committee members is not empty since once
-            // the committee is finalized, it should not be empty.
-            let responses = join_all(tasks).await;
-            let mut has_failure = false;
-            for response in responses {
-                if response.unwrap().effects.unwrap().status() != &SuiExecutionStatus::Success {
-                    has_failure = true;
-                }
-            }
-            if has_failure {
-                let bridge_summary = test_cluster.get_bridge_summary().await.unwrap();
-                assert_ne!(bridge_summary.committee.members.len(), 0);
-            }
-        }
-
-        info!(
-            "TestCluster build_with_bridge took {:?} secs",
-            timer.elapsed().as_secs()
-        );
-        test_cluster.bridge_authority_keys = Some(bridge_authority_keys);
-        test_cluster.bridge_server_ports = Some(server_ports);
-        test_cluster
     }
 
     /// Start a Swarm and set up WalletConfig
